requirements.txt
httplib2==0.9.1
python-instagram==1.3.1
simplejson==3.7.3
six==1.9.0
wsgiref==0.1.2
clarifai==0.2
googlemaps==2.2
factual-api==1.6.1
